Rejigger joseki UI a bit
This moves and styles things around a bit in an attempt to make it look
a bit more like how the puzzles or game page present the goban and
action/info.

This is a work in progress and not complete, but as far as I am aware
everything works as it did before, it just looks different.